+++ examples/lang-err-integration-tests/call-builder/lib.rs
@@ -54,8 +54,8 @@ mod call_builder {
             let result = build_call::<DefaultEnvironment>()
                 .call_type(Call::new().callee(address))
                 .exec_input(ExecutionInput::new(Selector::new(selector)))
-                .returns::<Result<(), ink::LangError>>()
-                .fire()
+                .returns::<()>()
+                .try_invoke()
                 .expect("Error from the Contracts pallet.");
 
             match result {
@@ -65,6 +65,24 @@ mod call_builder {
                     unimplemented!("No other `LangError` variants exist at the moment.")
                 }
             }
+        }
+
+        /// Call a contract using the `CallBuilder`.
+        ///
+        /// Since we can't use the `CallBuilder` in a test environment directly we need this
+        /// wrapper to test things like crafting calls with invalid selectors.
+        ///
+        /// This message does not allow the caller to handle any `LangErrors`, for that use the
+        /// `call` message instead.
+        #[ink(message)]
+        pub fn invoke(&mut self, address: AccountId, selector: [u8; 4]) {
+            use ink::env::call::build_call;
+
+            build_call::<DefaultEnvironment>()
+                .call_type(Call::new().callee(address))
+                .exec_input(ExecutionInput::new(Selector::new(selector)))
+                .returns::<()>()
+                .invoke()
         }
 
         /// Instantiate a contract using the `CreateBuilder`.
@@ -181,9 +199,7 @@ mod call_builder {
                 .call(&ink_e2e::alice(), flipper_get, 0, None)
                 .await
                 .expect("Calling `flipper::get` failed");
-            let initial_value = get_call_result
-                .return_value()
-                .expect("Input is valid, call must not fail.");
+            let initial_value = get_call_result.return_value();
 
             let selector = ink::selector_bytes!("invalid_selector");
             let call = build_message::<CallBuilderTestRef>(contract_acc_id)
@@ -193,9 +209,7 @@ mod call_builder {
                 .await
                 .expect("Calling `call_builder::call` failed");
 
-            let flipper_result = call_result
-                .return_value()
-                .expect("Call to `call_builder::call` failed");
+            let flipper_result = call_result.return_value();
 
             assert!(matches!(
                 flipper_result,
@@ -208,10 +222,52 @@ mod call_builder {
                 .call(&ink_e2e::alice(), flipper_get, 0, None)
                 .await
                 .expect("Calling `flipper::get` failed");
-            let flipped_value = get_call_result
-                .return_value()
-                .expect("Input is valid, call must not fail.");
+            let flipped_value = get_call_result.return_value();
             assert!(flipped_value == initial_value);
+
+            Ok(())
+        }
+
+        #[ink_e2e::test(additional_contracts = "../integration-flipper/Cargo.toml")]
+        async fn e2e_invalid_message_selector_panics_on_invoke(
+            mut client: ink_e2e::Client<C, E>,
+        ) -> E2EResult<()> {
+            let constructor = CallBuilderTestRef::new();
+            let contract_acc_id = client
+                .instantiate("call_builder", &ink_e2e::ferdie(), constructor, 0, None)
+                .await
+                .expect("instantiate failed")
+                .account_id;
+
+            let flipper_constructor = FlipperRef::new_default();
+            let flipper_acc_id = client
+                .instantiate(
+                    "integration_flipper",
+                    &ink_e2e::ferdie(),
+                    flipper_constructor,
+                    0,
+                    None,
+                )
+                .await
+                .expect("instantiate `flipper` failed")
+                .account_id;
+
+            // Since `LangError`s can't be handled by the `CallBuilder::invoke()` method we expect
+            // this to panic.
+            let invalid_selector = [0x00, 0x00, 0x00, 0x00];
+            let call = build_message::<CallBuilderTestRef>(contract_acc_id)
+                .call(|contract| contract.invoke(flipper_acc_id, invalid_selector));
+            let call_result = client.call(&ink_e2e::ferdie(), call, 0, None).await;
+
+            assert!(call_result.is_err());
+            let contains_err_msg = match call_result.unwrap_err() {
+                ink_e2e::Error::CallDryRun(dry_run) => {
+                    String::from_utf8_lossy(&dry_run.debug_message)
+                        .contains("Cross-contract call failed with CouldNotReadInput")
+                }
+                _ => false,
+            };
+            assert!(contains_err_msg);
 
             Ok(())
         }
@@ -243,8 +299,7 @@ mod call_builder {
                 .call(&ink_e2e::bob(), call, 0, None)
                 .await
                 .expect("Client failed to call `call_builder::call_instantiate`.")
-                .return_value()
-                .expect("Dispatching `call_builder::call_instantiate` failed.");
+                .return_value();
 
             assert!(
                 call_result.is_none(),
@@ -281,8 +336,7 @@ mod call_builder {
                 .call(&ink_e2e::charlie(), call, 0, None)
                 .await
                 .expect("Client failed to call `call_builder::call_instantiate`.")
-                .return_value()
-                .expect("Dispatching `call_builder::call_instantiate` failed.");
+                .return_value();
 
             assert!(
                 matches!(call_result, Some(ink::LangError::CouldNotReadInput)),
@@ -315,8 +369,8 @@ mod call_builder {
                 build_message::<CallBuilderTestRef>(contract_acc_id).call(|contract| {
                     contract.call_instantiate(code_hash, selector, init_value)
                 });
-            let call_result = client.call(&mut ink_e2e::dave(), call, 0, None).await;
 
+            let call_result = client.call(&mut ink_e2e::dave(), call, 0, None).await;
             assert!(
                 call_result.is_err(),
                 "Call execution should've failed, but didn't."
@@ -364,10 +418,10 @@ mod call_builder {
                 .await
                 .expect("Calling `call_builder::call_instantiate_fallible` failed")
                 .return_value()
-                .expect("Dispatching `call_builder::call_instantiate_fallible` failed.");
+                .expect("TODO");
 
             assert!(
-                call_result.unwrap().is_ok(),
+                call_result.is_ok(),
                 "Call to falliable constructor failed, when it should have succeeded."
             );
 
@@ -401,8 +455,7 @@ mod call_builder {
                 .call(&mut ink_e2e::ferdie(), call, 0, None)
                 .await
                 .expect("Calling `call_builder::call_instantiate_fallible` failed")
-                .return_value()
-                .expect("Dispatching `call_builder::call_instantiate_fallible` failed.");
+                .return_value();
 
             let contract_result = call_result
                 .unwrap()
@@ -493,8 +546,7 @@ mod call_builder {
                 .expect(
                     "Client failed to call `call_builder::call_instantiate_fallible`.",
                 )
-                .return_value()
-                .expect("Dispatching `call_builder::call_instantiate_fallible` failed.");
+                .return_value();
 
             assert!(
                 matches!(call_result, Some(Err(ink::LangError::CouldNotReadInput))),